requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==23.1
django-rest-framework-nested==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
PyJWT==2.6.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
